+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 12.1.2
 - Fix PHP warning when updating course period teacher in Moodle/functions.php & Moodle/Scheduling/Courses.php
 - Fix enrol teacher into course instead of just assigning teacher role in Moodle/functions.php, Moodle/Scheduling/Courses.php & Moodle/School_Setup/Rollover.php
 - MySQL fix You can't specify target table 'ac' for update in FROM clause in Courses.fnc.php
+- Fix Moodle error Short name is already used for another course in Moodle/Scheduling/Courses.php & Moodle/School_Setup/Rollover.php
 
 Changes in 12.1.1
 -----------------